pretrained/examples/PaddleHub/inference/image_classification/requirements.txt
paddlepaddle
paddlehub
shapely
pyclipper
opencv-python


